Verb tweaks (#31309)
* Verb tweaks

Remove the LOS check because this is already done above in CanExamine.

* Fix outlines

* import
